rename /typescript/ to /types/
