+++ lib/types/GlobalButterConfig.ts
